FIXED: Quintus foreign conversion routines: typing and exceptions

+++ src/SWI-Prolog.h
@@ -588,7 +588,7 @@ PL_EXPORT(int)		PL_cvt_i_float(term_t p, double *c);
 PL_EXPORT(int)		PL_cvt_i_single(term_t p, float *c);
 PL_EXPORT(int)		PL_cvt_i_string(term_t p, char **c);
 PL_EXPORT(int)		PL_cvt_i_atom(term_t p, atom_t *c);
-PL_EXPORT(int)		PL_cvt_o_integer(intptr_t c, term_t p);
+PL_EXPORT(int)		PL_cvt_o_integer(long c, term_t p);
 PL_EXPORT(int)		PL_cvt_o_float(double c, term_t p);
 PL_EXPORT(int)		PL_cvt_o_single(float c, term_t p);
 PL_EXPORT(int)		PL_cvt_o_string(const char *c, term_t p);
